Initial suggestion to break up the module by our friend Claude, with tiny modifications.
