+++ CHANGES.md
@@ -4,9 +4,14 @@ This is a record of all past dags releases and what went into them in reverse
 chronological order. We follow [semantic versioning](https://semver.org/) and all
 releases are available on [conda-forge](https://anaconda.org/conda-forge/dags).
 
-## 0.???
+## 0.3.0
 
-- :gh:`19` Update infrastructure, move to Pixi (:ghuser:`timmens`).
+- :gh:`29` Fix behavior of absolute/relative namespace handling (:ghuser:`hmgaudecker`)
+- :gh:`30` Refactor `dag_tree` (:ghuser:`hmgaudecker`)
+- :gh:`28` Expose relevant functions for working with function trees (:ghuser:`hmgaudecker`)
+- :gh:`26` Allow Unicode characters in Python identifiers (:ghuser:`MImmesberger`, :ghuser:`hmgaudecker`).
+- :gh:`24` Add type hints to the codebase (:ghuser:`timmens`, :ghuser:`hmgaudecker`).
+- :gh:`23` Update infrastructure, move to Pixi (:ghuser:`timmens`).
 - :gh:`17` Add possibility to use namespaces (:ghuser:`lars-reimann`, :ghuser:`MImmesberger`).
 
 ## 0.2.3
