+++ pyproject.toml
@@ -169,7 +169,9 @@ extend-ignore = [
 
 [tool.ruff.lint.per-file-ignores]
 "docs/source/conf.py" = ["INP001", "ERA001", "RUF100"]
+"src/dags/dag_tree/__init__.py" = ["RUF022"]
 "tests/*" = ["D401", "FBT001", "INP001", "PLC2401"]
+
 
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
